+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this programmer does her best to follow [Conventional Commits](https://www.c
 ### Changed
 
 - Moved PerfLogger instance listing utils to newly created Logger base class
+- Added jsonlines to dependancies
 
 ### Removed